# Simplify diffusion flux calls by removing need for sparse EB struct