--- a/Exec/RegTests/EB-C1/GNUmakefile
+++ b/Exec/RegTests/EB-C1/GNUmakefile
@@ -33,7 +33,7 @@ USE_MASA = TRUE
 ifeq ($(USE_MASA), TRUE)
   INCLUDE_LOCATIONS += ${MASA_HOME}/include ${MASA_HOME}/lib
   LIBRARY_LOCATIONS += ${MASA_HOME}/lib
-  LIBRARIES += -Wl,-rpath,${MASA_HOME}/lib -lmasa -lfmasa
+  LIBRARIES += -Wl,-rpath,${MASA_HOME}/lib -lmasa
   DEFINES += -DPELEC_USE_MASA
 endif