I think this is going to be one of many passes to start removing the sparse EB stuct stuff. 

As it stands:
- TG shows zero diffs with this:
```
fcompare.llvm.ex plt00010 orig/plt00010

            variable name            absolute error            relative error
                                        (||A - B||)         (||A - B||/||A||)
 ----------------------------------------------------------------------------
 level = 0
 density                                          0                         0
 Temp                                             0                         0
 rho_AIR                                          0                         0
 pressure                                         0                         0
 magvort                                          0                         0
 x_velocity                                       0                         0
 y_velocity                                       0                         0
 z_velocity                                       0                         0
 magvel                                           0                         0
 PLOTFILE AGREE
```

- EB-C10 shows zero diffs after 1 time step. But it starts to show diffs after 10:
```
fcompare.llvm.ex plt00010 orig/plt00010

            variable name            absolute error            relative error
                                        (||A - B||)         (||A - B||/||A||)
 ----------------------------------------------------------------------------
 level = 0
 density                            1.437034085e-13           1.221930097e-10
 xmom                               1.317701503e-11           6.074641266e-12
 ymom                                1.81393777e-10           9.289685792e-08
 zmom                               2.407705198e-10           1.234358553e-07
 rho_E                              5.681533366e-05           2.242645273e-11
 rho_e                              5.692942068e-05            2.24889264e-11
 Temp                               3.043726338e-08           1.014276429e-10
 rho_AIR                            1.437034085e-13           1.221930097e-10
 pressure                           2.277176827e-05            2.24889264e-11
 kineng                             1.138477046e-07           5.687605118e-11
 enstrophy                          9.230308251e-07           5.912866196e-10
 soundspeed                         1.761007297e-06           5.071420924e-11
 MachNumber                         2.647701502e-12           4.926423943e-11
 entropy                                          0                         0
 magvort                            9.377209835e-07            5.52749737e-10
 divu                               9.678583148e-07           3.413065229e-08
 eint_E                                0.2183890343           1.014275324e-10
 eint_e                                0.2183890343           1.014275324e-10
 logden                             5.398215208e-11           1.820239227e-11
 x_velocity                         1.567175332e-07           8.398970725e-11
 y_velocity                         1.568169946e-07           8.746746098e-08
 z_velocity                         2.081355642e-07           1.162141533e-07
 magvel                              1.56472197e-07           8.385821744e-11
 radvel                             7.023731996e-08           3.766137871e-11
 magmom                             1.319944154e-11           6.084979725e-12
 cp                                               0                         0
 cv                                               0                         0
 x                                                0                         0
 y                                                0                         0
 z                                                0                         0
 viscosity                                        0                         0
 bulk_viscosity                                   0                         0
 conductivity                                     0                         0
 D(AIR)                                           0                         0
 vfrac                                            0                         0
```
I am very confused as to why it doesn't diff after 1 step but it does on subsequent steps.

I am also a little worried about `pc_move_transcoefs_to_ec`. It's never been EB aware... should it be?!