@marchdf 2023-06-12
I am actually going to say I am ok with this. EB-C10 is a quiescent flow initially. EB-C1 which is an MMS case is much more interesting (non-zero fluxes). 100 steps with dev branch and this branch gives me machine diffs:
```            variable name            absolute error            relative error
                                        (||A - B||)         (||A - B||/||A||)
 ----------------------------------------------------------------------------
 level = 0
 density                             4.33680869e-19           3.367086133e-16
 Temp                               1.705302566e-13           5.143186602e-16
 rho_AIR                             4.33680869e-19           3.367086133e-16
 pressure                           8.149072528e-10            7.71988953e-16
 magvort                            1.455191523e-11            4.04060863e-16
 x_velocity                         1.818989404e-12           4.993697749e-16
 y_velocity                         1.818989404e-12           4.993697749e-16
 z_velocity                         1.818989404e-12           4.993697749e-16
 magvel                             1.818989404e-12           2.912749696e-16
 rhommserror                         4.33680869e-19             1.9535114e-15
 ummserror                          1.818989404e-12           3.516479422e-16
 vmmserror                          1.818989404e-12           3.516479422e-16
 wmmserror                          1.818989404e-12           3.516479422e-16
 pmmserror                           6.98491931e-10           1.639866137e-14
 vfrac                                            0                         0
```
So I am now ok with this code. Though I should probably check a multi-species case.