@marchdf 2023-06-15
I should have documented the speedup difference. I used the challenge problem with these additional CLI options: `max_step=100 amr.max_level=0 pelec.do_react=0 amr.checkpoint_files_output=0 amr.plot_files_output=0 amr.n_cell=128 128 32`. 

New code 37a33715 [full profile](https://github.com/AMReX-Combustion/PeleC/files/11761688/new.txt):

```-----------------------------------------------------------------------------------------------
Name                                            NCalls  Excl. Min  Excl. Avg  Excl. Max   Max %
-----------------------------------------------------------------------------------------------
PeleC::getMOLSrcTerm()                             200      24.85      24.85      24.85  42.17%
PeleC::extend_signed_distance()                     51      14.76      14.76      14.76  25.04%
PeleC::initialize_eb2_structs()                     51      11.74      11.74      11.74  19.92%
amrex::Copy()                                     5250      1.589      1.589      1.589   2.70%
PeleC::initialize_signed_distance()                 51      1.149      1.149      1.149   1.95%
PeleC::pc_apply_face_stencil()                     800      1.003      1.003      1.003   1.70%
StateData::FillBoundary(geom)                      804     0.9755     0.9755     0.9755   1.66%
Amr::readProbinFile()                                1     0.6451     0.6451     0.6451   1.09%
```

Old code 2ee36d71 [full profile](https://github.com/AMReX-Combustion/PeleC/files/11761700/old.txt)

```
-----------------------------------------------------------------------------------------------
Name                                            NCalls  Excl. Min  Excl. Avg  Excl. Max   Max %
-----------------------------------------------------------------------------------------------
PeleC::getMOLSrcTerm()                             200      26.68      26.68      26.68  44.11%
PeleC::extend_signed_distance()                     51      14.76      14.76      14.76  24.40%
PeleC::initialize_eb2_structs()                     51      11.77      11.77      11.77  19.45%
amrex::Copy()                                     5250      1.595      1.595      1.595   2.64%
PeleC::initialize_signed_distance()                 51      1.151      1.151      1.151   1.90%
PeleC::pc_apply_face_stencil()                     800      1.006      1.006      1.006   1.66%
StateData::FillBoundary(geom)                      804     0.8523     0.8523     0.8523   1.41%
Amr::readProbinFile()                                1     0.3918     0.3918     0.3918   0.65%
```

Almost a 7% speedup in `getMOLSrcTerm`!